Clippy pass, bump all crates to Rust 2024 edition

+++ dys-simulation/src/simulation/collision.rs
@@ -75,7 +75,7 @@ pub(crate) fn handle_collision_events(game_state: Arc<Mutex<GameState>>) -> Simu
 
                     let (rigid_body_set, _) = game_state.physics_sim.sets();
                     let combatant_rigid_body = rigid_body_set.get(combatant_rb).unwrap();
-                    combatant_rigid_body.linvel().magnitude() as f32
+                    combatant_rigid_body.linvel().magnitude()
                 };
 
                 let mut game_state = game_state.lock().unwrap();
